Demonstrate that 'defined' is better syntax
Coverage analysis performed on the 134441-Dumpvalue branch indicated
that the statement:

        print "empty slot\n";

inside unwrap() was not being reached.  Examination of the code
suggested that it was not being reached because the statement:

        exists $v->[$num]

in the 'if' condition would always return true.  Indeed, testing
demonstrates that what gets printed here is actually "undef\n" -- which
is not what one would expect on the basis of examination of the code.
What we would expect is "empty slot\n".

However, 'perldoc -f exists' strongly advises against using 'exists' on
array elements.  Since the 'slot' in the array always exists, we would
be better off saying:

        defined $v->[$num]

That's better Perl, and it changes what gets printed to "empty slot\n",
which probably was the original intent of the code -- but which is a
change of many years' behavior.